major updates to test
