--- .streamlit/config.toml
@@ -1,13 +1,0 @@
-[theme]
-base="light"
-font="sans serif"
-
-# Be sure to also set the theme to “Custom theme” in the hamburger menu
-# Sometimes it doesn’t work by simply placing the .toml file in the .streamlit folder.
-
-[logger]
-level = "info"
-
-# Turn off streamlit collecting summary statistics
-[browser]
-gatherUsageStats = false
