--- ACKNOWLEDGMENTS.md
@@ -1,20 +1,0 @@
-#### Acknowledgements
-
- **Overall Thanks**  
-Thank you to Syed Abbas http://github.com/syedabbast for showing me how to set up a WebApp using Streamlit and Github
-
-Thank you to the members of the U.S. Coast Guard Auxiliary for the encouragement and for giving essential feedback to tune ASK. Find out how you can help the Coast Guard by learning more about USCG Auxiliary [here](faqs.md) or [email me] (uscgaux.drew@wks.us).
-
-**Qdrant Vector Database**
-Qdrant has been a great tool because it is open source, works local, cloud and has a FREE TIER.  
-Qdrant Discord community [here](https://qdrant.to/discord) was great with answering quetions. Thank you @Anush008 and of course @andrey.vasnetsov
-
-**Streamlit**
-Thanks to the Streamlit Community for help developing a UI that was minimalistic but effective.
-
-**Embeddings and Inference**
-Deeplearning.ai has great [short courses](https://www.deeplearning.ai/short-courses/) in OpenI API prompting, Langchain and Weights & Biases instrumentation for how to build applications.
-
-## Credits
-- Background of github avatar Taylor Van Riper taylorvanripen925 under Unsplash image license
-- Suit in image Mary Ellen Matthews, NBC
