+++ admin/what_happens_when_i.py
@@ -1,9 +1,9 @@
 import streamlit as st 
 import datetime, time, os, sys
 from trubrics.integrations.streamlit import FeedbackCollector
-import ASK_inference as ASK
+import inference as ASK
 from qdrant_client import QdrantClient
-from ASK_inference import config
+from inference import config
 from streamlit_extras.stylable_container import stylable_container
