--- pages/library_overview.md
@@ -1,17 +1,0 @@
-
-ASK contains:
-- PDF documents. ASK includes only information that has been shared as a PDF.
-- Public Documents. All documents in ASK are available to the public.
-- National Sources: Documents have been curated from cgaux.org and uscg.mil
-- Documents issued by the USCG Auxiliary and the Coast Guard 
-- Policy documents
-
-ASK does NOT contain:
-- Private documents: ASK does not contain information from behind MemberZone, such as L2 Messages or most National Workshops
-- Outdated documents: ASK doesn't include cancelled, expired or superceded documents
-- Local or Regional documents
-- Information from AUXDATA or AUXOFFICER
-- ALAUX messages designated informational from past years, such as those commemorating/recognizing people or events
-- ALAUX promulgation announcements (which interfere with retrival of the policies themselves)
-- Biographies or points of contact lists
-- Event-related info such as schedules or calendars (e.g., C-School Calendar, NACON schedule)
